+++ src/ScatterPlot/ScatterPlotAxis.story.tsx
@@ -1,17 +1,9 @@
 import React, { Fragment, useState } from 'react';
-import { object, color, number } from '@storybook/addon-knobs';
-
 import { ScatterPlot } from './ScatterPlot';
 import {
-  signalChartData,
-  largeSignalChartData,
-  medSignalChartData,
-  signalStageData,
-  signalStages
+  largeSignalChartData
 } from '../../demo/signals';
 import { randomNumber, singleDateData } from '../../demo';
-import { range } from 'd3-array';
-import { GridlineSeries, Gridline, GridStripe } from '../common/Gridline';
 import { ScatterSeries, ScatterPoint } from './ScatterSeries';
 import {
   LinearYAxis,
@@ -21,142 +13,7 @@ import {
   LinearXAxisTickSeries,
   LinearXAxisTickLabel
 } from '../common/Axis/LinearAxis';
-import { symbolStar, symbol } from 'd3-shape';
-import { schemes } from '../common/color';
 import { getYScale, getXScale } from '../common/scales';
-
-export default {
-  title: 'Charts/Scatter Plot/Linear',
-};
-
-export const Simple = () => {
-    const height = number('Height', 400);
-    const width = number('Width', 750);
-    const size = number('Size', 4);
-    const fill = color('Color', schemes.cybertron[0]);
-    const data = object('Data', medSignalChartData);
-
-    return (
-      <ScatterPlot
-        height={height}
-        width={width}
-        data={data}
-        series={
-          <ScatterSeries point={<ScatterPoint color={fill} size={size} />} />
-        }
-      />
-    );
-  };
-
-export const CategoricalAxis = () => (
-    <ScatterPlot
-      height={400}
-      width={750}
-      data={signalStageData}
-      yAxis={
-        <LinearYAxis
-          type="category"
-          domain={signalStages as any}
-          tickSeries={
-            <LinearYAxisTickSeries
-              label={<LinearYAxisTickLabel rotation={false} />}
-            />
-          }
-        />
-      }
-      gridlines={
-        <GridlineSeries
-          line={<Gridline direction="y" />}
-          stripe={<GridStripe direction="y" />}
-        />
-      }
-    />
-  );
-
-export const NoAnimation = () => (
-    <ScatterPlot
-      height={400}
-      width={750}
-      data={medSignalChartData}
-      series={<ScatterSeries animated={false} />}
-    />
-  );
-
-export const Performance = () => (
-    <Fragment>
-      {range(15).map((i) => (
-        <div
-          key={i}
-          style={{
-            width: '250px',
-            height: '250px',
-            border: 'solid 1px green',
-            margin: '25px',
-            display: 'inline-block'
-          }}
-        >
-          <ScatterPlot data={medSignalChartData} />
-        </div>
-      ))}
-    </Fragment>
-  );
-
-export const Autosize = () => (
-    <div style={{ width: '50vw', height: '50vh', border: 'solid 1px red' }}>
-      <ScatterPlot data={medSignalChartData} />
-    </div>
-  );
-
-export const Symbols = () => (
-    <ScatterPlot
-      height={400}
-      width={750}
-      data={signalChartData}
-      series={
-        <ScatterSeries
-          point={
-            <ScatterPoint
-              symbol={() => {
-                const d = symbol().type(symbolStar).size(175)();
-
-                return (
-                  <path
-                    d={d!}
-                    style={{
-                      fill: 'lime',
-                      stroke: 'purple',
-                      strokeWidth: 1.5
-                    }}
-                  />
-                );
-              }}
-            />
-          }
-        />
-      }
-    />
-  );
-
-export const Bubble = () => (
-    <ScatterPlot
-      height={400}
-      width={750}
-      data={largeSignalChartData}
-      margins={20}
-      series={
-        <ScatterSeries
-          point={
-            <ScatterPoint
-              color="rgba(45, 96, 232, .8)"
-              size={(v) => v.metadata.severity + 5}
-            />
-          }
-        />
-      }
-    />
-  );
-
-export const LiveUpdate = () => <BubbleChartLiveUpdate />;
 
 export default {
   title: 'Charts/Scatter Plot/Axis',
